Web-Application/DeepDream/requirements.txt
Flask==2.3.2
numpy==1.24.3
Pillow==10.0.0
tensorflow==1.15.5